Fix Android build #rb trivial #rnx #jira none
#ROBOMERGE-SOURCE: CL 14535344 in //UE4/Release-4.26/...
#ROBOMERGE-BOT: RELEASE (Release-4.26 -> Release-Engine-Staging) (v749-14533626)

[CL 14535460 by jonas meyer in Release-Engine-Staging branch]

+++ Engine/Source/Runtime/VulkanRHI/Private/VulkanMemory.cpp
@@ -3830,7 +3830,7 @@ namespace VulkanRHI
 							return false;
 						}
 
-						SMoveAllocation M = { Index, AlignedOffset, AllocatedOffset, AllocatedSize } ;
+						SMoveAllocation M = { (uint32)Index, AlignedOffset, AllocatedOffset, AllocatedSize } ;
 						Moves.Add(M);
 					}
 					break;
